refactor protocol-related pathod methods
